% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateNotenlpCovariateSettings.R
\name{createNotenlpCovariateSettings}
\alias{createNotenlpCovariateSettings}
\title{createNotenlpCovariateSettings}
\usage{
createNotenlpCovariateSettings(
  useNotenlpData = TRUE,
  startDay = -30,
  endDay = 0,
  analysisId = 998,
  note_databaseschema = "cdm",
  note_tablename = "note",
  notenlp_databaseschema = "cdm",
  notenlp_tablename = "note_nlp",
  filter_select_domains = "",
  filter_include_nonexist = FALSE,
  filter_include_temporal = FALSE,
  filter_only_standard = FALSE,
  covariateDataSave = "",
  covariateDataLoad = ""
)
}
\description{
Create a covariateSettings object for constructing covariates from the note_nlp table in the OMOP CDM.
}
