% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLexicalVariants.R
\name{getLexicalVariants}
\alias{getLexicalVariants}
\title{getLexicalVariants}
\usage{
getLexicalVariants(
  covariateIds = NULL,
  conceptIds = NULL,
  connectionDetails,
  notenlp_database = "cdm",
  notenlp_table = "note_nlp"
)
}
\description{
Get the lexical variants of covariates(concepts) from the note_nlp table.
}
