% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTextRepresentationCovariates.R
\name{getTextRepCovariateData}
\alias{getTextRepCovariateData}
\title{getTextRepCovariateData}
\usage{
getTextRepCovariateData(
  connection,
  oracleTempSchema = NULL,
  cdmDatabaseSchema,
  cohortTable,
  cohortId,
  cdmVersion = "5",
  rowIdField = "subject_id",
  covariateSettings,
  aggregated = FALSE
)
}
\arguments{
\item{connection}{A connection to the server containing the schema as created using the connect function in the DatabaseConnector package.}

\item{oracleTempSchema}{A schema where temp tables can be created in Oracle.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM instance. Requires read permissions to this database. On SQL Server, this should specifiy both the database and the schema, so for example 'cdm_instance.dbo'.}

\item{cohortTable}{Name of the table holding the cohort for which we want to construct covariates. If it is a temp table, the name should have a hash prefix, e.g. '#temp_table'. If it is a non-temp table, it should include the database schema, e.g. 'cdm_database.cohort'.}

\item{cohortId}{For which cohort ID should covariates be constructed? If set to -1, covariates will be constructed for all cohorts in the specified cohort table.}

\item{cdmVersion}{The version of the Common Data Model used. Currently only cdmVersion = "5" is supported.}

\item{rowIdField}{The name of the field in the cohort temp table that is to be used as the row_id field in the output table. This can be especially usefull if there is more than one period per person.}

\item{covariateSettings}{An object of type covariateSettings as created using the createTextRepCovariateSettings function.}

\item{aggregated}{Should aggregate statistics be computed instead of covariates per cohort entry?}
}
\value{
a covariateData object that can be used my other cdm framework functions.
}
\description{
This covariate builder creates text representation covariates from a cohort in the OMOP cdm
}
